refactor: update functions directory where used
